Add habbts to database

+++ app/src/main/java/skills/future/planer/ui/tasklist/TaskListCreatorFragment.java
@@ -64,7 +64,7 @@ public class TaskListCreatorFragment extends Fragment {
 
         processFabColor();
 
-        Integer taskID = -1;
+        Long taskID = -1L;
         taskID = getTaskIDFromFragmentArgument(taskID);
         saveBtnOnClickListenerSetter(taskID);
 
@@ -119,9 +119,9 @@ public class TaskListCreatorFragment extends Fragment {
     /**
      * Checks is edit of task or create new task
      */
-    private Integer getTaskIDFromFragmentArgument(Integer taskID) {
+    private Long getTaskIDFromFragmentArgument(Long taskID) {
         if (getArguments() != null) {
-            taskID = (Integer) getArguments().get("idTaskToEdit");
+            taskID = (Long) getArguments().get("idTaskToEdit");
             if (taskID != -1)
                 loadDataFromTask(taskID);
         }
@@ -157,7 +157,7 @@ public class TaskListCreatorFragment extends Fragment {
     /**
      * Gets TaskData from database, and sets fields of creatorFragment
      */
-    private void loadDataFromTask(Integer taskID) {
+    private void loadDataFromTask(Long taskID) {
         taskDataDao = AppDatabase.getInstance(this.getContext()).taskDataTabDao();
         try {
             editTask = taskDataDao.findById(taskID);
@@ -209,7 +209,7 @@ public class TaskListCreatorFragment extends Fragment {
      * But if task is edited
      * listener sends edited TaskData
      */
-    private void saveBtnOnClickListenerSetter(Integer taskID) {
+    private void saveBtnOnClickListenerSetter(Long taskID) {
         saveButton.setOnClickListener(view1 -> {
             AnimateView.animateInOut(saveButton, getContext());
 
@@ -248,7 +248,7 @@ public class TaskListCreatorFragment extends Fragment {
     /**
      * Sends task to database depending on the creator type
      */
-    private void sendTaskToDataBase(Integer taskID, View view1) {
+    private void sendTaskToDataBase(Long taskID, View view1) {
         //If the task is edited
         if (taskID != -1)
             taskDataDao.editOne(editTask);
